pyproject.toml
[project]
name = "jcutil"
version = "2.0.1"
description = "A tool chain for the web service"
readme = "README.md"
license = {text = "MIT"}
authors = [
    { name = "Jochen.he", email = "thjl@hotmail.com" }
]
requires-python = ">=3.8"
dependencies = [
    "apscheduler>=3.11.0",
    "colorama>=0.4.6",
    "jcramda>=1.0.6",
    "joblib>=1.4.2",
    "motor>=3.5.1",
    "pandas>=2.0.3",
    "py-consul>=1.6.0",
    "pycryptodomex>=3.22.0",
    "pyhcl>=0.4.5",
    "pymongo>=4.10.1",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "redis>=5.2.1",
    "tornado>=6.3.3",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "anyio>=4.5.2",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.11.5",
]

[tool.ruff]
# 启用的规则集
lint.select = [
    "E",  # pycodestyle 错误
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "W",  # pycodestyle 警告
]

# 忽略特定规则
lint.ignore = [
    "E501",  # 行过长警告
]

# 行长度设置
line-length = 100

# 目标Python版本
target-version = "py38"

# 自动修复设置
[tool.ruff.format]
quote-style = "single"
indent-style = "space"
line-ending = "auto"

# 导入设置
[tool.ruff.lint.isort]
known-first-party = ["jcutil"]
